feat: ui updates

+++ src/components/DeleteCastButton.tsx
@@ -2,9 +2,10 @@
 
 import useLocalStorage from "@/hooks/use-local-storage-state";
 import { UserInfo } from "@/types";
-import { useRouter } from "next/navigation";
 import { useState, type FC } from "react";
 import { toast } from "react-toastify";
+import { Button } from "./ui/button";
+import { Cross1Icon } from "@radix-ui/react-icons";
 
 interface Props {
   hash: string;
@@ -13,9 +14,14 @@ interface Props {
 
 export const DeleteCastButton: FC<Props> = ({ hash, author }) => {
   const [user] = useLocalStorage<UserInfo>("user");
-  const router = useRouter();
+  const [loading, setLoading] = useState(false);
 
   const deleteCast = async () => {
+    if (!user.signerUuid) {
+      return;
+    }
+
+    setLoading(true);
     try {
       const req = await fetch("/api/casts", {
         method: "DELETE",
@@ -28,8 +34,6 @@ export const DeleteCastButton: FC<Props> = ({ hash, author }) => {
         }),
       });
 
-      const res = await req.json();
-
       if (req.ok) {
         toast("Cast deleted successfully!", {
           type: "success",
@@ -37,7 +41,7 @@ export const DeleteCastButton: FC<Props> = ({ hash, author }) => {
           position: "bottom-right",
         });
 
-        router.push(`/profile/${user.fid}`);
+        window.location.reload();
       } else {
         toast("Error deleting cast", {
           type: "error",
@@ -46,12 +50,13 @@ export const DeleteCastButton: FC<Props> = ({ hash, author }) => {
         });
       }
     } catch (e) {
-      console.error(e);
       toast("Error deleting cast", {
         type: "error",
         autoClose: 5000,
         position: "bottom-right",
       });
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -60,16 +65,21 @@ export const DeleteCastButton: FC<Props> = ({ hash, author }) => {
   return (
     <>
       {Number(user.fid) === Number(author) && (
-        <button
+        <Button
+          variant="ghost"
           onClick={deleteCast}
-          style={{
-            backgroundColor: "red",
-            color: "white",
-            marginTop: "1rem",
-          }}
+          className="flex items-center ml-auto hover:bg-transparent hover:text-red-500"
+          disabled={loading}
         >
-          Delete a cast
-        </button>
+          {loading ? (
+            "Deleting..."
+          ) : (
+            <>
+              Remove
+              <Cross1Icon className="w-4 h-4 ml-4" />
+            </>
+          )}
+        </Button>
       )}
     </>
   );
